Fix an enourmous amount of alignment-related bugs
To aid with this, a new hybrid-API for making use of x86's AC-flag was added, however use of it is quite complicated due to pretty much the entire windows-API expecting unaligned memory accesses to be allowed.
Anyways, deemon has just become much more capable when it comes to running on ARM

+++ src/deemon/objects/unicode/format.c
@@ -751,9 +751,9 @@ check_print_error:
   self->f_flush_start = self->f_iter;
  }
  /* Flush the remainder. */
- return (*printer)(arg,self->f_flush_start,
-                  (size_t)(self->f_iter -
-                           self->f_flush_start));
+ return (int)(*printer)(arg,self->f_flush_start,
+                       (size_t)(self->f_iter -
+                                self->f_flush_start));
 #undef print
 err_arg:
  Dee_Decref(in_arg);
